refactor(ef): partner-manage-party 핸들러 분리 (827→53 lines) (#1873)
index.ts 827 lines → 53 lines thin router로 분리.
action별 핸들러를 _handlers/, 공유 상수·검증기를 _lib/으로 추출.

기존 통합 테스트 44건 + 핸들러별 단위 테스트 20건 전부 통과.

Closes #1873

Constraint: EF 단일 진입점(index.ts) 원칙 유지 — 라우팅만 담당
Rejected: handleRequest 함수 유지 | 불필요한 래퍼 레이어 제거로 index.ts 더 단순화
Confidence: high
Scope-risk: narrow
